# perf: update benchmarks to record throughput